refactor: Relocate `useMediaUpload` to the editor component
This hook is used solely within this component at this time.
